+++ rpc/backend/tracing.go
@@ -26,6 +26,7 @@ import (
 	rpctypes "github.com/evmos/ethermint/rpc/types"
 	ethermint "github.com/evmos/ethermint/types"
 	evmtypes "github.com/evmos/ethermint/x/evm/types"
+	feemarkettypes "github.com/evmos/ethermint/x/feemarket/types"
 	"github.com/pkg/errors"
 )
 
@@ -117,7 +118,13 @@ func (b *Backend) TraceTransaction(hash common.Hash, config *rpctypes.TraceConfi
 		// 0 is a special value in `ContextWithHeight`
 		contextHeight = 1
 	}
-	traceResult, err := b.queryClient.TraceTx(rpctypes.ContextWithHeight(contextHeight), &traceTxRequest)
+	// Get basefee from transaction height
+	res, err := b.queryClient.FeeMarket.Params(rpctypes.ContextWithHeight(transaction.Height), &feemarkettypes.QueryParamsRequest{})
+	if err != nil {
+		return nil, err
+	}
+	ctx := rpctypes.ContextWithBaseFee(rpctypes.ContextWithHeight(contextHeight), res.Params.BaseFee.String())
+	traceResult, err := b.queryClient.TraceTx(ctx, &traceTxRequest)
 	if err != nil {
 		return nil, err
 	}